move parallel module to linear.py and fix comments
- add del_trainer as a classmethod to ParallelOVRTrainer to ensure a proper close
